configParser add file lock, fix restart system when rental app halted

+++ lib/util.py
@@ -12,7 +12,7 @@ import datetime
 
 import logger
 import codecs
-import msvcrt
+# import msvcrt
 
 log = logger.Logger().get()
 
@@ -45,12 +45,14 @@ def update_config(ini_file, set_item={}, section='SYNC'):
             conf.readfp(codecs.open(ini_file, "r", "utf-8-sig"))
 
             if conf.has_section(section):
+                """update section items"""
                 for key in set_item:
                     if conf.has_option(section, key):
                         conf.set(section, key, set_item[key])
             else:
-                log.error('section:'+str(section)+'not found, copy '+str(ini_file)+'.sample')
                 """copy the sample file as ini file"""
+                log.error('section:'+str(section)+'not found, copy '+str(ini_file)+'.sample')
+
                 from shutil import copyfile
                 sample_conf=ini_file+'.sample'
                 """unlink origin ini"""
@@ -59,27 +61,44 @@ def update_config(ini_file, set_item={}, section='SYNC'):
 
             fsize=os.path.getsize(ini_file)
 
-            config_file=open(ini_file, 'w+')
-            """on windows, write lock"""
-            fd=config_file.fileno()
-            msvcrt.locking(fd, msvcrt.LK_RLCK, fsize)
-            conf.write(config_file)
-            config_file.close()
-            msvcrt.locking(fd, msvcrt.LK_UNLCK, fsize)
+            """use plain file to emulate lock"""
+            lock_file=os.sep.join([os.getcwd(), 'sync.lock'])
 
-            """on windows endline with \r\n"""
-            f = open(ini_file, 'r+')
-            msvcrt.locking(f.fileno(), msvcrt.LK_RLCK, fsize)
-            """r'\n' raw string"""
-            conf_data=f.read()
-            if len(conf_data) > 0:
-                conf_data=conf_data.replace(r'\n', r'\r\n')
-                f.seek(0)
-                f.write(conf_data)
+            if not os.path.exists(lock_file):
+                try:
+                    flock=open(lock_file, 'w')
 
-            f.close()
-            msvcrt.locking(f.fileno(), msvcrt.LK_UNLCK, fsize)
-            return True
+                    config_file=open(ini_file, 'w+')
+                    """on windows, write lock"""
+                    # fd=config_file.fileno()
+                    # msvcrt.locking(fd, msvcrt.LK_RLCK, fsize)
+                    conf.write(config_file)
+                    # msvcrt.locking(fd, msvcrt.LK_UNLCK, fsize)
+                    config_file.close()
+
+                    """on windows endline with \r\n"""
+                    f = open(ini_file, 'r+')
+                    # msvcrt.locking(f.fileno(), msvcrt.LK_RLCK, fsize)
+                    """r'\n' raw string"""
+                    conf_data=f.read()
+                    if len(conf_data) > 0:
+                        conf_data=conf_data.replace(r'\n', r'\r\n')
+                        f.seek(0)
+                        f.write(conf_data)
+
+                    # msvcrt.locking(f.fileno(), msvcrt.LK_UNLCK, fsize)
+                    f.close()
+
+                    flock.close()
+                    os.remove(lock_file)
+
+                    return True
+
+                except Exception, e:
+                    log.error(str(e))
+
+            else:
+                return False
 
         except Exception, e:
             log.error(str(e))
